# docs/swagger.yaml
...
openapi: 3.1.0
info:
  title: Six Can Do Eat API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  "/api/v1/areas/{areaId}":
    get:
      tags:
        - Area
      summary: 운영 지역 상세 조회
      operationId: getArea
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SuccessResponseAreaResponse"
    put:
      tags:
        - Area
      summary: 운영 지역 수정
      operationId: updateArea
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateAreaRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SuccessResponseAreaResponse"
    delete:
      tags:
        - Area
      summary: 운영 지역 삭제 소프트 삭제)
      operationId: deleteArea
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SuccessResponseVoid"
...
  "/api/v1/areas":
    get:
      tags:
        - Area
      summary: 운영 지역 목록 조회
      operationId: getAreas
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: createdAt
        - name: isAsc
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SuccessResponsePageAreaResponse"
    post:
      tags:
        - Area
      summary: 운영 지역 등록
      operationId: createArea
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AreaCreateRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SuccessResponseAreaResponse"
...
components:
  schemas:
...
    UpdateAreaRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        city:
          type: string
          maxLength: 50
          minLength: 0
        district:
          type: string
          maxLength: 50
          minLength: 0
        is_active:
          type: boolean
    AreaResponse:
      type: object
      properties:
        areaId:
          type: string
          format: uuid
        name:
          type: string
        city:
          type: string
        district:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        is_active:
          type: boolean
    SuccessResponseAreaResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          "$ref": "#/components/schemas/AreaResponse"
...
    AreaCreateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        city:
          type: string
          maxLength: 50
          minLength: 0
        district:
          type: string
          maxLength: 50
          minLength: 0
      required:
        - city
        - district
        - name
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          "$ref": "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    PageAreaResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/AreaResponse"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        empty:
          type: boolean
    SuccessResponsePageAreaResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          "$ref": "#/components/schemas/PageAreaResponse"
    SuccessResponseVoid:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT